# demo/synthetic_servers/v1/costco/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Costco",
    "version": "1.0.0",
    "description": "API for Costco wholesale shopping service"
  },
  "paths": {
    "/api/v1/products": {
      "get": {
        "summary": "Get products with optional category filter",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/membership": {
      "get": {
        "summary": "Get membership details",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Membership details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders": {
      "get": {
        "summary": "Get user's order history",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Place new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created"
          }
        }
      }
    },
    "/api/v1/warehouses": {
      "get": {
        "summary": "Get nearby warehouses",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of warehouses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Warehouse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "name": {"type": "string"},
          "category": {"type": "string"},
          "price": {"type": "number"},
          "item_number": {"type": "string"},
          "description": {"type": "string"},
          "in_stock": {"type": "boolean"},
          "is_member_only": {"type": "boolean"}
        }
      },
      "Membership": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "type": {"type": "string"},
          "status": {"type": "string"},
          "expiration_date": {"type": "string"},
          "member_since": {"type": "string"},
          "auto_renewal": {"type": "boolean"}
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "total": {"type": "number"},
          "warehouse_id": {"type": "string"},
          "order_date": {"type": "string"},
          "status": {"type": "string"}
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "product_id": {"type": "string"},
          "quantity": {"type": "integer"},
          "price": {"type": "number"}
        }
      },
      "Warehouse": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "name": {"type": "string"},
          "address": {"type": "string"},
          "phone": {"type": "string"},
          "hours": {"type": "string"},
          "latitude": {"type": "number"},
          "longitude": {"type": "number"}
        }
      },
      "NewOrder": {
        "type": "object",
        "properties": {
          "user_email": {"type": "string"},
          "warehouse_id": {"type": "string"},
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      }
    }
  }
}